+++ include/chen/data/ini.hpp
@@ -30,15 +30,16 @@
 #include <string>
 #include <chen/base/utf8.hpp>
 #include <chen/base/str.hpp>
+#include <chen/base/num.hpp>
 
 namespace chen
 {
     class ini
     {
     public:
-        typedef std::unordered_map<std::string, std::string> property_type;
-        typedef std::unordered_map<std::string, property_type> value_type;
-        typedef std::pair<std::string, property_type> section_type;
+        typedef std::unordered_map<std::string, std::string> property_type;  // k/v pair
+        typedef std::pair<std::string, property_type> section_type;          // each section
+        typedef std::unordered_map<std::string, property_type> value_type;   // multiple sections
 
         class error : public std::runtime_error
         {
@@ -52,54 +53,120 @@ namespace chen
 
     public:
         /**
-         * Ini parse and stringify
+         * Ini parse, accept text, file or even iterators
          */
         static chen::ini::value_type parse(const std::string &text, bool file = false);
-        static std::string stringify(const chen::ini::value_type &map);
+
+        template <typename InputIterator>
+        static chen::ini::value_type parse(InputIterator cur, InputIterator end);
 
         /**
-         * Decode ini, throw exception if found error
+         * Ini stringify
          */
-        template <typename InputIterator>
-        static chen::ini::value_type decode(InputIterator cur, InputIterator end);
+        static std::string stringify(const chen::ini::value_type &map);
 
     protected:
         /**
          * Throw syntax exception
          */
         template <typename InputIterator>
-        static void exception(InputIterator &cur, InputIterator &end);
+        static void exception(const InputIterator &beg, InputIterator &cur, InputIterator &end);
+
+        /**
+         * Advance to the next non-whitespace character
+         */
+        template <typename InputIterator>
+        static bool advance(const InputIterator &beg, InputIterator &cur, InputIterator &end, bool require = true);
 
         /**
          * Decode specific type
          */
         template <typename InputIterator>
-        static chen::ini::section_type decodeSection(InputIterator &cur, InputIterator &end);
+        static void decode(chen::ini::value_type &out, const InputIterator &beg, InputIterator &cur, InputIterator &end);
 
         template <typename InputIterator>
-        static chen::ini::property_type decodeProperty(InputIterator &cur, InputIterator &end);
+        static void decode(chen::ini::section_type &out, const InputIterator &beg, InputIterator &cur, InputIterator &end);
 
         template <typename InputIterator>
-        static std::string decodeString(InputIterator &cur, InputIterator &end);
+        static void decode(chen::ini::property_type &out, const InputIterator &beg, InputIterator &cur, InputIterator &end);
 
         template <typename InputIterator>
-        static void decodeComment(InputIterator &cur, InputIterator &end);
+        static void decode(std::string &out, const InputIterator &beg, InputIterator &cur, InputIterator &end);
+
+        template <typename InputIterator>
+        static void comment(const InputIterator &beg, InputIterator &cur, InputIterator &end);
     };
+}
+
+// parse
+template <typename InputIterator>
+chen::ini::value_type chen::ini::parse(InputIterator cur, InputIterator end)
+{
+    chen::ini::value_type item;
+
+    const InputIterator beg = cur;
+
+    // trim left spaces
+    if (!chen::ini::advance(beg, cur, end, false))
+        return item;
+
+    // decode item
+    chen::ini::decode(item, beg, cur, end);
+
+    // trim right spaces
+    chen::ini::advance(beg, cur, end, false);
+
+    // should reach end
+    if (cur != end)
+        chen::ini::exception(beg, cur, end);
+
+    return item;
+}
+
+// exception
+template <typename InputIterator>
+void chen::ini::exception(const InputIterator &beg, InputIterator &cur, InputIterator &end)
+{
+    if (cur == end)
+    {
+        throw error("ini: unexpected end of input");
+    }
+    else
+    {
+        auto pos = chen::num::str(std::distance(beg, cur));
+        throw error(chen::str::format("ini: unexpected token '%c' at position %s", *cur, pos.c_str()));
+    }
+}
+
+// advance
+template <typename InputIterator>
+bool chen::ini::advance(const InputIterator &beg, InputIterator &cur, InputIterator &end, bool require)
+{
+    // skip whitespaces
+    while ((cur != end) && std::isspace(*cur))
+        ++cur;
+
+    // check if end
+    if (cur == end)
+    {
+        if (require)
+            chen::ini::exception(beg, cur, end);
+        else
+            return false;
+    }
+
+    return true;
 }
 
 // decode
 template <typename InputIterator>
-chen::ini::value_type chen::ini::decode(InputIterator cur, InputIterator end)
+void chen::ini::decode(chen::ini::value_type &out, const InputIterator &beg, InputIterator &cur, InputIterator &end)
 {
-    chen::ini::value_type value;
     bool root = true;
 
     while (cur != end)
     {
-        for (; (cur != end) && std::isspace(*cur); ++cur)
-            ;
-
-        if (cur == end)
+        if (!chen::ini::advance(beg, cur, end, false))
             break;
 
         switch (*cur)
@@ -107,50 +174,42 @@ chen::ini::value_type chen::ini::decode(InputIterator cur, InputIterator end)
             case '[':
             {
                 root = false;
-                auto section = chen::ini::decodeSection(cur, end);
 
-                if (value.find(section.first) == value.end())
-                    value.emplace(std::move(section));
+                chen::ini::section_type s;
+                chen::ini::decode(s, beg, cur, end);
+
+                if (out.find(s.first) == out.end())
+                    out.emplace(std::move(s));
                 else
-                    chen::ini::exception(cur, end);
+                    chen::ini::exception(beg, cur, end);
             }
                 break;
 
             case ';':
-                chen::ini::decodeComment(cur, end);
+                chen::ini::comment(beg, cur, end);
                 break;
 
             default:
                 if (root)
-                    value.emplace("", std::move(chen::ini::decodeProperty(cur, end)));
+                {
+                    chen::ini::property_type p;
+                    chen::ini::decode(p, beg, cur, end);
+                    out.emplace("", std::move(p));
+                }
                 else
-                    chen::ini::exception(cur, end);
+                {
+                    chen::ini::exception(beg, cur, end);
+                }
                 break;
         }
     }
-
-    return value;
 }
 
-// exception
 template <typename InputIterator>
-void chen::ini::exception(InputIterator &cur, InputIterator &end)
+void chen::ini::decode(chen::ini::section_type &out, const InputIterator &beg, InputIterator &cur, InputIterator &end)
 {
-    // todo report line and column
-    if (cur == end)
-        throw error("ini: unexpected end of input");
-    else
-        throw error(str::format("ini: unexpected token '%c'", *cur));
-}
-
-// decode
-template <typename InputIterator>
-chen::ini::section_type chen::ini::decodeSection(InputIterator &cur, InputIterator &end)
-{
-    chen::ini::section_type section;
-
     if ((cur == end) || (*cur != '['))
-        chen::ini::exception(cur, end);
+        chen::ini::exception(beg, cur, end);
     else
         ++cur;
 
@@ -161,33 +220,27 @@ chen::ini::section_type chen::ini::decodeSection(InputIterator &cur, InputIterat
         name += *cur++;
 
     if (name.empty() || (cur == end) || (*cur != ']'))
-        chen::ini::exception(cur, end);
+        chen::ini::exception(beg, cur, end);
     else
         ++cur;
 
-    section.first = std::move(name);
+    out.first = std::move(name);
 
     // properties
-    for (; (cur != end) && std::isspace(*cur); ++cur)
-        ;
+    chen::ini::advance(beg, cur, end, false);
 
-    section.second = std::move(chen::ini::decodeProperty(cur, end));
-
-    return section;
+    chen::ini::property_type p;
+    chen::ini::decode(p, beg, cur, end);
+    out.second = std::move(p);
 }
 
 template <typename InputIterator>
-chen::ini::property_type chen::ini::decodeProperty(InputIterator &cur, InputIterator &end)
+void chen::ini::decode(chen::ini::property_type &out, const InputIterator &beg, InputIterator &cur, InputIterator &end)
 {
-    chen::ini::property_type property;
-
     while (cur != end)
     {
         // skip
-        for (; (cur != end) && std::isspace(*cur); ++cur)
-            ;
-
-        if ((cur == end) || (*cur == '['))
+        if (!chen::ini::advance(beg, cur, end, false) || (*cur == '['))
             break;
 
         std::string key;
@@ -200,11 +253,11 @@ chen::ini::property_type chen::ini::decodeProperty(InputIterator &cur, InputIter
         str::trim(key);
 
         if (key.empty())
-            chen::ini::exception(cur, end);
+            chen::ini::exception(beg, cur, end);
 
         // equal
         if ((cur == end) || (*cur != '='))
-            chen::ini::exception(cur, end);
+            chen::ini::exception(beg, cur, end);
         else
             ++cur;
 
@@ -213,33 +266,38 @@ chen::ini::property_type chen::ini::decodeProperty(InputIterator &cur, InputIter
             ;
 
         if ((cur != end) && (*cur != '\n'))
-            val = std::move(chen::ini::decodeString(cur, end));
+        {
+            std::string s;
+            chen::ini::decode(s, beg, cur, end);
+            val = std::move(s);
+        }
 
         // store
-        if (property.find(key) == property.end())
-            property.emplace(std::move(key), std::move(val));
+        if (out.find(key) == out.end())
+        {
+            out.emplace(std::move(key), std::move(val));
+        }
         else
-            throw error("ini: duplicate key found: " + key);
+        {
+            auto pos = chen::num::str(std::distance(beg, cur) - 4);
+            throw error(chen::str::format("ini: duplicate key %s found at position %s", key.c_str(), pos.c_str()));
+        }
 
         // skip
         if ((cur != end) && (*cur == '\n'))
             ++cur;
     }
-
-    return property;
 }
 
 template <typename InputIterator>
-std::string chen::ini::decodeString(InputIterator &cur, InputIterator &end)
+void chen::ini::decode(std::string &out, const InputIterator &beg, InputIterator &cur, InputIterator &end)
 {
-    std::string ret;
+    if ((cur == end) || std::isspace(*cur))
+        chen::ini::exception(beg, cur, end);
+
     bool quote = false;
 
-    if ((cur == end) || std::isspace(*cur))
-    {
-        chen::ini::exception(cur, end);
-    }
-    else if (*cur == '"')
+    if (*cur == '"')
     {
         quote = true;
         ++cur;
@@ -249,63 +307,63 @@ std::string chen::ini::decodeString(InputIterator &cur, InputIterator &end)
     {
         if (!quote && (*cur == ';'))
         {
-            chen::ini::decodeComment(cur, end);
+            chen::ini::comment(beg, cur, end);
             break;
         }
         else if (*cur != '\\')
         {
-            ret += *cur++;
+            out += *cur++;
             continue;
         }
 
         // escape
         if (++cur == end)
-            chen::ini::exception(cur, end);
+            chen::ini::exception(beg, cur, end);
 
         switch (*cur)
         {
             case '\\':
-                ret += '\\';
+                out += '\\';
                 break;
 
             case '0':
-                ret += '\0';
+                out += '\0';
                 break;
 
             case 'a':
-                ret += '\a';
+                out += '\a';
                 break;
 
             case 'b':
-                ret += '\b';
+                out += '\b';
                 break;
 
             case 't':
-                ret += '\t';
+                out += '\t';
                 break;
 
             case 'r':
-                ret += '\r';
+                out += '\r';
                 break;
 
             case 'n':
-                ret += '\n';
+                out += '\n';
                 break;
 
             case ';':
-                ret += ';';
+                out += ';';
                 break;
 
             case '#':
-                ret += '#';
+                out += '#';
                 break;
 
             case ':':
-                ret += ':';
+                out += ':';
                 break;
 
             case '"':
-                ret += '\"';
+                out += '\"';
                 break;
 
             case 'x':
@@ -313,36 +371,38 @@ std::string chen::ini::decodeString(InputIterator &cur, InputIterator &end)
                 ++cur;
 
                 // handle unicode character
-                char unicode[5] = {0};
+                char unicode[5];
+                unicode[4] = '\0';
 
                 for (auto i = 0; i < 4; ++i, ++cur)
                 {
                     if (cur == end)
-                        chen::ini::exception(cur, end);
+                        chen::ini::exception(beg, cur, end);
 
                     auto ch = *cur;
 
                     if (((ch >= '0') && (ch <= '9')) || ((ch >= 'a') && (ch <= 'f')) || ((ch >= 'A') && (ch <= 'F')))
                         unicode[i] = ch;  // char must in range of '0' ~ '9', 'a' ~ 'f', 'A' ~ 'F'
                     else
-                        chen::ini::exception(cur, end);
+                        chen::ini::exception(beg, cur, end);
                 }
 
                 // convert utf-16 to utf-8
                 try
                 {
-                    ret.append(chen::utf8::convert(static_cast<std::uint32_t>(std::strtol(unicode, nullptr, 16))));
+                    out.append(chen::utf8::convert(static_cast<std::uint32_t>(std::strtol(unicode, nullptr, 16))));
                 }
                 catch (...)
                 {
                     // e.g: \xD83D\xDE00, it's a emoji character
-                    throw error(str::format("ini: invalid unicode char \\x%s", unicode));
+                    auto pos = chen::num::str(std::distance(beg, cur) - 4);
+                    throw error(chen::str::format("ini: invalid unicode char \\u%s at position %s", unicode, pos.c_str()));
                 }
             }
                 break;
 
             default:
-                chen::ini::exception(cur, end);
+                chen::ini::exception(beg, cur, end);
                 break;
         }
 
@@ -352,24 +412,22 @@ std::string chen::ini::decodeString(InputIterator &cur, InputIterator &end)
     if (quote)
     {
         if ((cur == end) || (*cur != '"'))
-            chen::ini::exception(cur, end);
+            chen::ini::exception(beg, cur, end);
         else
             ++cur;
     }
     else
     {
-        str::rtrim(ret);
+        str::rtrim(out);
     }
-
-    return ret;
 }
 
 template <typename InputIterator>
-void chen::ini::decodeComment(InputIterator &cur, InputIterator &end)
+void chen::ini::comment(const InputIterator &beg, InputIterator &cur, InputIterator &end)
 {
-    if (*cur != ';')
-        chen::ini::exception(cur, end);
+    if ((cur == end) || (*cur != ';'))
+        chen::ini::exception(beg, cur, end);
 
-    for (; (cur != end) && (*cur != '\n'); ++cur)
-        ;
+    while ((cur != end) && (*cur != '\n'))
+        ++cur;
 }